Update combined_recieve_and_execute.ino
converted to using seconds for check_step() (don't even think about using minutes, unsigned long divided by an integer as large as 60 * 1000 doesn't work and breaking it into steps is more work than it is worth)

Also gives a readout of exactly what you're printing back to python during the send_and_check phase b/c at one point the Arduino just kept failing to send back what we gave to it and it turned out I tried to send a step duration value in ms and it was so large it wrapped back around as an int, which ended up as a negative value. 

Also tells when food reward is dispensed.